+++ im-server/src/main/java/vip/qsos/im/repository/db/IChatGroupOfLastRecordRepository.kt
@@ -3,4 +3,4 @@ package vip.qsos.im.repository.db
 import org.springframework.data.jpa.repository.JpaRepository
 import vip.qsos.im.model.db.TableChatGroupOfLastRecord
 
-interface IChatGroupOfLastRecordRepository : JpaRepository<TableChatGroupOfLastRecord, Int>
+interface IChatGroupOfLastRecordRepository : JpaRepository<TableChatGroupOfLastRecord, Long>
