+++ im-server/src/main/java/vip/qsos/im/repository/db/IChatGroupRepository.kt
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import vip.qsos.im.model.db.TableChatGroup
 import vip.qsos.im.model.type.ChatType
 
-interface IChatGroupRepository : JpaRepository<TableChatGroup, Int> {
+interface IChatGroupRepository : JpaRepository<TableChatGroup, Long> {
 
     fun findByMember(member: String): TableChatGroup?
     fun findByMemberLike(member: String): List<TableChatGroup>
